#92 Fix get updates on

+++ src/main/java/dev/tobee/telegram/request/body/SendDocumentBody.java
@@ -3,7 +3,7 @@ package dev.tobee.telegram.request.body;
 import java.nio.file.Path;
 import java.util.List;
 import java.util.Optional;
-import java.util.OptionalInt;
+import java.util.OptionalLong;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import dev.tobee.telegram.model.media.InputFile;
@@ -21,7 +21,7 @@ public record SendDocumentBody(
         @JsonProperty("disable_content_type_detection") Optional<Boolean> disableContentTypeDetection,
         @JsonProperty("disable_notification") Optional<Boolean> disableNotification,
         @JsonProperty("protect_content") Optional<Boolean> protectContent,
-        @JsonProperty("reply_to_message_id") OptionalInt replyToMessageId,
+        @JsonProperty("reply_to_message_id") OptionalLong replyToMessageId,
         @JsonProperty("allow_sending_without_reply") Optional<Boolean> allowSendingWithoutReply,
         @JsonProperty("reply_markup") Optional<ReplyMarkup> replyMarkup
 ) {
