+++ src/main/java/dev/tobee/telegram/request/body/SendInvoiceBody.java
@@ -3,6 +3,7 @@ package dev.tobee.telegram.request.body;
 import java.util.List;
 import java.util.Optional;
 import java.util.OptionalInt;
+import java.util.OptionalLong;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import dev.tobee.telegram.model.LabeledPrice;
@@ -32,7 +33,7 @@ public record SendInvoiceBody(
         @JsonProperty("is_flexible") Optional<Boolean> isFlexible,
         @JsonProperty("disable_notification") Optional<Boolean> disableNotification,
         @JsonProperty("protect_content") Optional<Boolean> protectContent,
-        @JsonProperty("reply_to_message_id") OptionalInt replyToMessageId,
+        @JsonProperty("reply_to_message_id") OptionalLong replyToMessageId,
         @JsonProperty("allow_sending_without_reply") Optional<Boolean> allowSendingWithoutReply,
         @JsonProperty("reply_markup") Optional<ReplyMarkup> replyMarkup
 ) {
